Updated the DispatchMessages fluent API so that the default overload returns the IObservable.
This was done to make sure that we don't hide IObservable because in doing so we limit what the end user can do further downstream with Rx APIs. For example, we want to handle exceptions from message handler with Rx's .Catch(...) API.

There are still extension methods to allow for "simpler" configurations that just want to supply an IObserver or individual action methods.